0.003
This doesn't work yet! I'm slowly converting the example WDM AC97 driver to work for HD Audio and I'm only as far as the start of common.cpp. Builds with MSVC++ 6.0, can debug with WinDbg over a (real or virtualized) serial port. Code for reference is in the "BleskOS Sound Driver" and "Warp Code Generated" folders.