Try to have ichwave manage the BDL
this is a dead end. there's no way to control the alignment of the buffers we get from the kernel mixer (and i have no way of invalidating mappings without stopping the stream) so I need to re-implement this around WaveCyclic instead. i thought that might be the case but this confirms it.