BANG YOUR HEAD 🔊💀🤘
audio will drive ya mad. Got *something* playing out of the speakers in this new (somewhat simpler) framework. It's still just popping the speaker continuously with no attention paid to the volume slider or anything that's supposed to be written into the buffer. But I can fix SOME of that immediately and at least this structure is MORE understandable to make changes to.