It works! I can't believe it!
More smaller buffers helped immensely. System sounds are now coming through OK, but Winamp is still making horrible screeching noises. Need to implement sample rate changing because I think it's expecting 22050 even if I set it to 44100. Do audio buffer sizes need to be power of two? I don't get anything with 1792-byte buffers. Still a lot of bugs, I would not try this without the kernel debugger attached because if it fails an assert it's going to freeze.