Check STATESTS for codec attach
At least get an interrupt and do a debug print. Not sure when exactly the IRQ fires on a real codec but on VMWare by the time it fired we were already halfway through codec initialization.
I want to see on real hardware i915 and i945 if there's ever an attach interrupt or if there's still something we need to do in ACPI or BIOS or something to power up the Codec and have it respond.